# ConnectedCarApi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API's",
    "description": "Connected Car API's",
    "version": "1.0"
  },
  "produces": ["application/json"],  
  "host": "connected-car-api.mybluemix.net",
  "basePath": "/",
  "paths": {
...
	  "/auth/facebook/token": {
		  "post": {
			"x-swagger-router-controller": "social-auth",
			"operationId": "facebookLogin",
			"tags": ["Auth"],
			"description": "Authenticate the user by facebook token.",   
			"parameters": [
			  {
				"name": "access_token",
				"in": "header",
				"description": "Access token",
				"required": true,
				"type": "string"
			  }
			],		
			"responses": {
			  "default": {
				"description": "Invalid request.",
				"schema": {
				  "$ref": "#/definitions/Error"
				}
			  },
			  "200": {
				"description": "Successful request.",
				"schema": {
				  "type": "array",
				  "$ref": "#/definitions/Token"
				}
			  }
			}
		  }
	  },
	  "/auth/google/token": {
		  "post": {
			"x-swagger-router-controller": "social-auth",
			"operationId": "googleLogin",
			"tags": ["Auth"],
			"description": "Authenticate the user by google token.",   
			"parameters": [
			  {
				"name": "access_token",
				"in": "header",
				"description": "Access token",
				"required": true,
				"type": "string"
			  }
			],		
			"responses": {
			  "default": {
				"description": "Invalid request.",
				"schema": {
				  "$ref": "#/definitions/Error"
				}
			  },
			  "200": {
				"description": "Successful request.",
				"schema": {
				  "type": "array",
				  "$ref": "#/definitions/Token"
				}
			  }
			}
		  }
	  },
	  "/auth/google/token/id": {
		  "post": {
			"x-swagger-router-controller": "social-auth",
			"operationId": "googleIdLogin",
			"tags": ["Auth"],
			"description": "Authenticate the user by google id token.",   
			"parameters": [
			  {
				"name": "access_token",
				"in": "header",
				"description": "Access token",
				"required": true,
				"type": "string"
			  }
			],		
			"responses": {
			  "default": {
				"description": "Invalid request.",
				"schema": {
				  "$ref": "#/definitions/Error"
				}
			  },
			  "200": {
				"description": "Successful request.",
				"schema": {
				  "type": "array",
				  "$ref": "#/definitions/Token"
				}
			  }
			}
		  }
	  },
...
  },
  "definitions": {
	"Token": {
		"properties": {
			"token": {
				"type": "string"
			},
			"expires": {
				"type": "number"
			},
			"user": {
				"$ref": "#/definitions/User" 
			}
		}
	},
...
    "User": {
	  "properties": {
        "username": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      }
    },
...
	"Error": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    },
...
  }
}